Add xbox_teleop_full.py file
